# docs/openapi.yaml
openapi: 3.0.0
info:
  title: IAM Service
  description: |
    IAM Service (Identity and Access Management Service) - это микросервис, который отвечает за управление 
    идентификацией и доступом пользователей
  version: 0.0.0
servers:
  - url: http://localhost:18081
    description: Local Server URL
...
paths:
...
  /api/v0/iam/register/confirm:
    get:
      tags:
        - Registration
      summary: Подтверждение регистрации пользователя
      description: Проверяет токен подтверждения и при успехе сохраняет пользователя
      operationId: confirmRegistration
      parameters:
        - name: token
          in: query
          description: Токен подтверждения
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
        '400':
          description: Токен подтверждения неверен или просрочен
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                Токен подтверждения пропущен:
                  description: Токен подтверждения пропущен
                  value:
                    created_at: '2024-03-27T03:26:19.385Z'
                    message: 'The confirmation token must be not null.'
                Токен подтверждения неверен или просрочен:
                  description: Токен подтверждения неверен или просрочен
                  value:
                    created_at: '2024-03-27T03:26:19.385Z'
                    message: 'The confirmation token is invalid or has expired.'
        '409':
          description: Имя пользователя или адрес электронной почты заняты
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                created_at: '2024-03-27T03:26:19.385Z'
                message: Username already exists.
        '500':
          description: Ошибка на сервере
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                created_at: '2024-03-27T03:26:19.385Z'
                message: Internal server error.
...
components:
  schemas:
    ErrorDto:
      required:
        - created_at
        - message
      type: object
      properties:
        message:
          type: string
          description: Сообщение ошибки
          example: 'Validation error:'
        errors:
          type: object
          additionalProperties:
            type: string
            description: Название поля и его ошибки
            example: '{"username":"The username must start with a letter and contain only Latin letters, numbers and underscores."}'
          description: Название поля и его ошибки
          example:
            username: The username must start with a letter and contain only Latin letters, numbers and underscores.
        created_at:
          type: string
          description: Время создания
          format: date-time
          example: '2024-03-27T03:57:26.951Z'
...