# src/main/java/com/archipio/iamservice/controller/api/v0/RegistrationController.java
package com.archipio.iamservice.controller.api.v0;

import static com.archipio.iamservice.util.PathUtils.API_V0_PREFIX;
import static com.archipio.iamservice.util.PathUtils.REGISTRATION_CONFIRM_SUFFIX;
import static com.archipio.iamservice.util.PathUtils.REGISTRATION_SUFFIX;
import static org.springframework.http.HttpStatus.ACCEPTED;
import static org.springframework.http.HttpStatus.CREATED;

import com.archipio.iamservice.dto.RegistrationDto;
import com.archipio.iamservice.exception.NullConfirmationTokenException;
import com.archipio.iamservice.service.RegistrationService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequiredArgsConstructor
@RequestMapping(API_V0_PREFIX)
public class RegistrationController {

  private final RegistrationService registrationService;

  @PostMapping(REGISTRATION_SUFFIX)
  public ResponseEntity<Void> register(@Valid @RequestBody RegistrationDto registrationDto) {
    registrationService.register(registrationDto);
    return ResponseEntity.status(ACCEPTED).build();
  }

  @GetMapping(REGISTRATION_CONFIRM_SUFFIX)
  public ResponseEntity<Void> confirmRegistration(@RequestParam("token") String token) {
    if (token == null || token.isEmpty()) {
      throw new NullConfirmationTokenException();
    }
    registrationService.confirmRegistration(token);
    return ResponseEntity.status(CREATED).build();
  }
}